% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NMAdata}
\alias{NMAdata}
\title{R pakage NMAnalysis by Hao Li, Ming-Hui Chen}
\format{A data frame with 73 rows and 15 variables}
\description{
This file is part of the R package NMAnalysis
}
\details{
LDL-C Network Meta-data for Analysis


The dataset is also listed in the supplement material of the submitted 
manuscript
}
\references{
H. Li, M-H. Chen, J.G. Ibrahim, S. Kim, A.K. Shah, 
J. Lin & A.M. Tershakovec (2017).
Bayesian Inference for Network Meta-Regression Using 
Multivariate Random Effects with Applications to Cholesterol Lowering Drugs.
UConn, Department of Statistics, Technical Report 17-05.
}
