% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMAnorm.R
\name{NMAmodel}
\alias{NMAmodel}
\title{Bayesian network meta-analysis}
\usage{
NMAmodel(ID_model, Nrep = 100L, Nthinning = 1L, Nwarmup = 10L,
  Iseed = 1234567L, alpha = 0.05, ID_study = NMAdata[, 1],
  ID_arm = NMAdata[, 2], Npt = NMAdata[, 3], Y = NMAdata[, 4],
  SD = NMAdata[, 5], X = matrix(as.double(as.matrix(NMAdata[, 6:15])), nrow
  = 73, ncol = 10), DIC = rep(0, 5), logCPO = rep(0, 29),
  betapo = matrix(0, 10, 4), gampo = matrix(0, 11, 4), tau2po = matrix(0,
  11, 4), eRho = matrix(0, 11, 11), seRho = matrix(0, 11, 11),
  Rholow = matrix(0, 11, 11), Rhoupp = matrix(0, 11, 11))
}
\arguments{
\item{ID_model}{model ID}

\item{Nrep}{MCMC sample size}

\item{Nthinning}{thinning size}

\item{Nwarmup}{warm-up size}

\item{Iseed}{random seed for the FORTRAN random number generator}

\item{alpha}{desired HPD credible level}

\item{ID_study}{study ID}

\item{ID_arm}{arm ID}

\item{Npt}{vector of length N for the number of patients}

\item{Y}{vector of length N for the response variable (aggregate mean)}

\item{SD}{vector of length N for the standard deviation}

\item{X}{design matrix, N by p}

\item{DIC}{(Output) a vector of length 5}

\item{logCPO}{(Output) a vector of length K, K is the number of trials}

\item{betapo}{(Output) a matrix of nx by 4, nx is the number of covariates}

\item{gampo}{(Output) a matrix of nT by 4, nT is the number of treatments}

\item{tau2po}{(Output) a matrix of nT by 4}

\item{eRho}{(Output) a matrix of nT by nT}

\item{seRho}{(Output) a matrix of nT by nT}

\item{Rholow}{(Output) a matrix of nT by nT}

\item{Rhoupp}{(Output) a matrix of nT by nT}
}
\value{
Bayesian comparison criteria (DIC and LPML), posterior
estimate of coefficients of covariates, treatment effects, variances
and correlation matrix of treatment random effects
}
\description{
This function links in the NMAmodel.f which builds
the network meta-regression proposed by Li et al. (2017)
}
